move uniswap client into pkg
