+++ api/server.go
@@ -18,13 +18,13 @@ import (
 	"github.com/vultisig/vultisigner/internal/syncer"
 	"github.com/vultisig/vultisigner/internal/tasks"
 	"github.com/vultisig/vultisigner/internal/types"
+	"github.com/vultisig/vultisigner/pkg/uniswap"
 	"github.com/vultisig/vultisigner/plugin"
 	"github.com/vultisig/vultisigner/plugin/dca"
 	"github.com/vultisig/vultisigner/plugin/payroll"
 	"github.com/vultisig/vultisigner/service"
 	"github.com/vultisig/vultisigner/storage"
 	"github.com/vultisig/vultisigner/storage/postgres"
-	"github.com/vultisig/vultisigner/uniswap"
 
 	"github.com/DataDog/datadog-go/statsd"
 	gcommon "github.com/ethereum/go-ethereum/common"
@@ -659,4 +659,3 @@ func (s *Server) VerifyCode(c echo.Context) error {
 	}
 	return c.NoContent(http.StatusOK)
 }
-
